apps/api-server/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "jinwooseok",email = "jinus7949@naver.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.127.1,<0.128.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "dependency-injector (>=4.48.3,<5.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "minio (>=7.2.20,<8.0.0)",
    "pylint (>=4.0.4,<5.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "aiomysql (>=0.3.2,<0.4.0)",
    "apscheduler (>=3.11.2,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "asgi-lifespan (>=2.1.0,<3.0.0)",
    "pytest-json-report (>=1.5.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "prometheus-fastapi-instrumentator (>=7.1.0,<8.0.0)",
    "boto3 (>=1.42.59,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
